FizzBuzz/FizzBuzz.cs
namespace FizzBuzzNS
{
    public class FizzBuzz
    {
        public string FizzBuzzMethod(int value)
        {
            if (value % 3 == 0 && value % 5 == 0)
            {
                return "FizzBuzz";
            }
            else if (value % 3 == 0)
            {
                return "Fizz";
            }
            else if (value % 5 == 0)
            {
                return "Buzz";
            }
            else
            {
                return value.ToString();
            }
        }
    }
}
